Apply readelf patch for cross-compiling on older Python versions (#452)
Linux cross-compiles are failing on `main` @
https://github.com/astral-sh/python-build-standalone/commit/01cfc3b2dff416f2390cb6d45faa67a384abc644
due to `readelf` missing. We drop this check as part of the Apple cross
compilation patch but since #443 we no longer apply that patch on Linux.
This moves the patch out into a separate file and applies it when
cross-compiling on Python <=3.12 where it is relevant.