Upgrade LLVM to 19.1.6 (#462)
This is a redo of #420, which was merged prematurely.

With the BOLT changes from #463 merged, LLVM 19 _just works_.

As part of this we also modernize the BOLT apply settings to follow the
recommendations at
https://llvm.org/devmtg/2024-03/slides/practical-use-of-bolt.pdf.

This includes enabling support for loading hot code from a huge page at
runtime. This should _just work_ and could result in perf wins via
improved iTLB hit rate, etc.