Implement matrix sharding (#492)
Since we've encountered the GitHub matrix size limit of 256 entries a
couple times now (recently in #388), this adds sharding to our matrix
generation script.

There's now a `--max-shards` option which splits the matrix into (up to)
the requested number of shards. We don't split it evenly, we eagerly
fill each shard first. To consume the shards, we just copy the `build`
job into two identical child jobs which consume each matrix shard.
Ideally, we'd avoid code duplication here by using a re-usable action or
something for the build-steps — but that's overkill at this point.

We only need this for Linux for the foreseeable future.