Cache dependency downloads on Linux (#566)
In https://github.com/astral-sh/python-build-standalone/pull/565 we note
we're seeing rate-limiting on downloads of bzip2 — which is blocking
development here. There, we increase the exponential backoff but we need
to reduce the number of downloads to address the root cause of the
issue. @geofft suggested caching these downloads, which makes a lot of
sense for us — there's already logic to validate the file if it exists
on disk.

I'm just starting with Linux.