Compile sqlite with FTS3 enhanced query syntax enabled (#562)
Fixes #550

This adds three compilation flags to the embedded sqlite module:

- `-DSQLITE_ENABLE_FTS3` to explicitly enable FTS3 support (which AFAIK
is already enabled by default)
- `-DSQLITE_ENABLE_FTS3_PARENTHESIS` to enable the enhanced FTS3 query
syntax (which isn't enabled by default)
- `-DSQLITE_ENABLE_FTS3_TOKENIZER` to enable the two-argument version of
the `fts3_tokenizer()` interface. It still has to be enabled on the
connection-level using `Connection.setconfig`

All those flags are enabled by default in the Debian build of SQLite,
and probably in other distributions as well.