fix(workflow): use correct input variable for release-create step (#648)
The release-create step was using an incorrect variable name
(`github.events.input.tag`), which resulted in the `just release-create`
command being called without the required tag argument. This commit
updates the workflow to use the correct variable
(`github.event.inputs.tag`), ensuring the tag is properly passed to the
command.

-
https://github.com/loong64/python-build-standalone/actions/runs/15469452121/job/43549720438

Signed-off-by: 吴小白 <296015668@qq.com>