Fix download statistics (#706)
`just download-stats*` are failing with an HTTP 504 Gateway Timeout on
GitHub's servers.

We can work around this by reducing the requested number of releases per
page.

As part of this, we adopt a 3rd party package for querying the GitHub
API because I was too lazy to implement pagination manually. Also, since
pagination is enabled, the counts reported by these commands are now
accurate and don't implicitly truncate at the most recent 10 releases.

Co-authored-by: Zanie Blue <contact@zanie.dev>