Build zstd with multithreading enabled (#762)
I verified this on macOS

```
❯ python -c "from compression import zstd; print(zstd.CompressionParameter.nb_workers.bounds())"
(0, 256)
```

See #761 

I'm not sure if this is all of the performance difference, but using
threads should make a significant difference.
https://github.com/astral-sh/python-build-standalone/issues/761#issuecomment-3193682600
notes that threads are not used by default, so I think there's more
going on — but we should enable this at build time regardless. See also
#768.