Replace `setup-just` with `setup-crate` (#771)
The former is just a composite of a latter, but does not pin to a commit
so it violates our policy. I do not think `setup-crate` supports pinning
a just commit, so I think we will want to just drop the whole action in
the future.

See 
- https://github.com/astral-sh/python-build-standalone/issues/760
- https://github.com/extractions/setup-just/issues/20
- https://github.com/extractions/setup-just/issues/23

See failure at
https://github.com/astral-sh/python-build-standalone/actions/runs/17052072811/attempts/1